+++ src/main/java/com/onlol/fetcher/repository/ChampionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChampionRepository extends JpaRepository<Champion, Integer> {
-    Champion findByChampId(Integer id);
+public interface ChampionRepository extends JpaRepository<Champion, Long> {
+    Champion findByChampId(Long id);
 }
